Fix:Click events on the Team Resources conversion page
