+++ src/pages/AddCluster/ResourceConversion/index.less
@@ -7,10 +7,11 @@
 	float:left;
 	width: 150px;
 	margin-right: 24px;
+	position: relative;
     .fixed{
-        position: fixed;
-        top: 180px;
-        left:100px;
+        position: absolute;
+        top: 35px;
+        left:-70px;
     }
 	.typeBtn {
 		display: block;
@@ -61,6 +62,10 @@
             }
         }
     }
+	.buttonbox{
+		background-color: #fff;
+		border: 1px solid #e8e8e8;
+	}
 	.ramstyle{
 		margin-left: 32px;
 		display: flex;
